Add manual page.
Just adding a manual page here isn't enough for the system to recognise it.
For a man page to be properly installed it needs to be on the $MANPATH (or in
some other directory that can be arrived at from $PATH using $MANPATH_MAP).

For the purposes of the talk, I added project's bin directory to my $PATH,
which meant that the project's man directory could be found using the
default $MANPATH_MAP.

See `man man` for all the gory details.